Initial Moth Jump Implementation
Imported moth jump audio files and added them to the file structure.

Created a new AudioStreamPlayer2D beneath the player and attached SFX.

Created new player animations:
1. moth jump start
2. moth jump hold
3. moth jump release

Created a new method inside the player audio manager to handle the release of the moth jump:
_on_moth_jump_released

Added Method Call tracks to the above animations to trigger the appropriate sound effects.

As with the other animations, this is untested. I am still figuring out the easiest way to test this sort of thing.